+++ app/controllers/companies/configurations_controller.rb
@@ -29,7 +29,7 @@ module Companies
     end
 
     def configuration_params
-      params.require(:jsonb_columns_configuration).permit(:work_start_time, :work_end_time)
+      params.require(:jsonb_columns_configuration).permit(:work_start_time, :work_end_time, insight_fields: {})
     end
   end
 end